\name{coef.tbl}
\alias{coef.tbl}
\title{Function to extract coefficient table from lmer object}
\usage{
  coef.tbl(fm, ...)
}
\arguments{
  \item{fm}{lmer object}
}
\description{
  Input lmer object output coefficient table
}
\details{
  Extracts coefficients from lmer object for use to check
  adequacy of simulation.
}

