% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{compute_power}
\alias{compute_power}
\title{Tidy Power Simulation}
\usage{
compute_power(sim_args)
}
\arguments{
\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{...}{Other arguments to pass to error simulation functions.}
}
\description{
Tidy Power Simulation
}
\examples{

}
