% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_variables.r
\name{corr_variables}
\alias{corr_variables}
\title{Function to correlate variables}
\usage{
corr_variables(mat, cor_vars, cov_param, standardize = TRUE)
}
\arguments{
\item{mat}{A matrix of variables to correlate}

\item{cor_vars}{A vector of correlations to specify, must be specified by row
where the first element is the correlation between variable 1 and variable 2, 
second correlation is between variable 1 and variable 3, and so on.}

\item{cov_param}{Variable specification similar to specifying fixed effects. 
See \code{\link{sim_reg}} for more details.}

\item{standardize}{TRUE/FALSE flag indicating whether variables should be standardized
prior to correlating (this is needed for accurate correlated variables)}
}
\description{
Inputs a matrix and other parameters and outputs a correlated matrix
}
