% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataReg.r
\name{data.reg.nested3}
\alias{data.reg.nested3}
\title{Simulates three level nested data with a single third level random effect}
\usage{
data.reg.nested3(Xmat, Zmat, Zmat3, beta, rand.eff, rand.eff3, k, n, p, err)
}
\arguments{
\item{Xmat}{A matrix of covariates.}

\item{Zmat}{Design matrix for random effects.}

\item{Zmat3}{Design matrix for level 3 random effects.}

\item{beta}{A vector of regression parameters.}

\item{rand.eff}{A vector of random effects, must be stacked.}

\item{rand.eff3}{A vector of level 3 random effects, must be stacked.}

\item{k}{Number of third level clusters.}

\item{n}{Number of clusters.}

\item{p}{Number of units within each cluster.}

\item{err}{A vector of within cluster errors.}
}
\description{
Takes simulation parameter arguments and
returns simulated data.
}

