% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glm.r
\name{data_glm_nested}
\alias{data_glm_nested}
\title{Generate logistic regression outcome}
\usage{
data_glm_nested(Xmat, Zmat, beta, rand_eff, n, p)
}
\arguments{
\item{Xmat}{A matrix of covariates.}

\item{Zmat}{Design matrix for random effects.}

\item{beta}{A vector of regression parameters.}

\item{rand_eff}{A vector of random effects, must be stacked.}

\item{n}{Number of clusters.}

\item{p}{Number of units within each cluster.}
}
\description{
Takes simulation parameter arguments and 
returns simulated data in the logistic metric and
converted to 0/1 based on the probabilities and the 
binomial distribution with 1 trial.
}
