% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reg.r
\name{data_reg_nested}
\alias{data_reg_nested}
\title{Simulates two level nested data}
\usage{
data_reg_nested(Xmat, Zmat, beta, rand_eff, n, p, err)
}
\arguments{
\item{Xmat}{A matrix of covariates.}

\item{Zmat}{Design matrix for random effects.}

\item{beta}{A vector of regression parameters.}

\item{rand_eff}{A vector of random effects, must be stacked.}

\item{n}{Number of clusters.}

\item{p}{Number of units within each cluster.}

\item{err}{A vector of within cluster errors.}
}
\description{
Takes simulation parameter arguments and 
returns simulated data.
}

