% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reg.r
\name{data_reg_single}
\alias{data_reg_single}
\title{Simulates single level data}
\usage{
data_reg_single(Xmat, beta, n, err)
}
\arguments{
\item{Xmat}{A matrix of covariates.}

\item{beta}{A vector of regression parameters.}

\item{n}{Number of clusters.}

\item{err}{A vector of within cluster errors.}
}
\description{
Takes simulation parameter arguments and returns simulated data.
}
\details{
This is a helper function to the master function \code{\link{sim_reg}}, 
this function does the actual simulation to return the data for single 
level models.
}

