% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extract Coefficients}
\usage{
extract_coefficients(model, extract_function = NULL)
}
\arguments{
\item{model}{A returned model object from a fitted model.}

\item{extract_function}{A function that extracts model results. The 
function must take the model object as the only argument.}
}
\description{
Extract Coefficients
}
