% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.r
\name{fit_propensity}
\alias{fit_propensity}
\title{Primary propensity model fitting}
\usage{
fit_propensity(data, sim_args)
}
\arguments{
\item{data}{Data object that contains the simulated data.}

\item{sim_args}{The simulation argument list that contains the propensity
score arguments needed.}
}
\value{
Propensity probabilities and weights if specified.
}
\description{
Incorporates the estimated probabilities and weights.
}
