% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.r
\name{generate_missing}
\alias{generate_missing}
\title{Tidy Missing Data Function}
\usage{
generate_missing(data, sim_args)
}
\arguments{
\item{data}{Data simulated from other functions to pass to this function.}

\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}
}
\description{
Tidy Missing Data Function
}
