% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reg.r
\name{generate_response}
\alias{generate_response}
\title{Simulate response variable}
\usage{
generate_response(data, sim_args, keep_intermediate = TRUE, ...)
}
\arguments{
\item{data}{Data simulated from other functions to pass to this function.}

\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{keep_intermediate}{TRUE/FALSE flag indicating whether intermediate steps
should be kept. This would include fixed effects times regression weights,
random effect summations, etc. Default is TRUE.}

\item{...}{Other arguments to pass to error simulation functions.}
}
\description{
Simulate response variable
}
