% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{HPD}
\alias{HPD}
\title{Internal function for finding highest probability density intervals from
a vector of simulations}
\usage{
HPD(x, probs, side)
}
\arguments{
\item{x}{vector of simulation values}

\item{probs}{numeric specifying the probability interval}

\item{side}{character either 'upper' or 'lower' specifying which side of the
interval to return.}
}
\description{
Internal function for finding highest probability density intervals from
a vector of simulations
}

