% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missing_data.r
\name{missing_data}
\alias{missing_data}
\title{Master Missing Data Function}
\usage{
missing_data(sim_data, resp_var = "sim.data", clust_var = "clustID",
  miss_prop, type = c("dropout", "random"))
}
\arguments{
\item{sim_data}{Simulated data frame}

\item{resp_var}{Response variable to add missing data to}

\item{clust_var}{Cluster variable used for the grouping.}

\item{miss_prop}{Proportion of missing data overall or a vector
the same length as the number of clusters representing the
percentage of missing data for each cluster}

\item{type}{The type of missing data to generate, currently supports
droput or random missing data.}
}
\description{
This function is a wrapper to easily call the specific types of
missing data mechanisms.
}

