% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{model_fit}
\alias{model_fit}
\title{Tidy Model Fitting Function}
\usage{
model_fit(data, sim_args, formula = NULL, model_function = NULL, ...)
}
\arguments{
\item{data}{A data object, most likely generated from within simglm}

\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{formula}{A formula for model fitting. If NULL, the default, the 
same formula for data simulation in used for model fitting.}

\item{model_function}{A model function. By default is 'lm' with no random 
effect and is 'lmer' with random effects. Able to specify any model
fitting function.}

\item{...}{Additional arguments needed to pass on to model fitting 
functions. See specific model fitting functions for specific details.}
}
\description{
Tidy Model Fitting Function
}
\examples{

}
