% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.r
\name{parse_crossclass}
\alias{parse_crossclass}
\title{Parse Cross-classified Random Effects}
\usage{
parse_crossclass(sim_args, random_formula_parsed)
}
\arguments{
\item{sim_args}{Simulation arguments}

\item{random_formula_parsed}{This is the output from 
\code{\link{parse_randomeffect}}.}
}
\description{
Parse Cross-classified Random Effects
}
