% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.r
\name{parse_fixed_type}
\alias{parse_fixed_type}
\title{Parses type of variable generation fixed effects}
\usage{
parse_fixed_type(fixed_parameters)
}
\description{
Parses type of variable generation fixed effects
}
