% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.r
\name{parse_multiplemember}
\alias{parse_multiplemember}
\title{Parse Multiple Membership Random Effects}
\usage{
parse_multiplemember(sim_args, random_formula_parsed)
}
\arguments{
\item{sim_args}{Simulation arguments}

\item{random_formula_parsed}{This is the output from 
\code{\link{parse_randomeffect}}.}
}
\description{
Parse Multiple Membership Random Effects
}
