% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.r
\name{parse_random}
\alias{parse_random}
\title{Parses random effect specification}
\usage{
parse_random(formula)
}
\arguments{
\item{formula}{Random effect formula already parsed by \code{\link{parse_formula}}}
}
\description{
Parses random effect specification
}
