% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.r
\name{parse_varyarguments_w}
\alias{parse_varyarguments_w}
\title{Parse within varying arguments}
\usage{
parse_varyarguments_w(sim_args, name)
}
\arguments{
\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{name}{The name of the within simulation condition. This is primarily
an internal function.}
}
\description{
Parse within varying arguments
}
