% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbimod.r
\name{rbimod}
\alias{rbimod}
\title{Simulating mixture normal distributions}
\usage{
rbimod(n, mean, var, num_dist)
}
\arguments{
\item{n}{Number of random draws.  Optionally can be a vector with number 
in each simulated normal distribution.}

\item{mean}{Vector of mean values for each normal distribution.
Must be the same length as num_dist.}

\item{var}{Vector of variance values for each normal distribution.
Must be the same length as num_dist.}

\item{num_dist}{Number of normal distributions to use when simulating mixture
normal distribution.}
}
\description{
Input simulation metrics returns mixture normal random variable.
}
\details{
Function to simulate mixture normal distributions.  The function computes
adds the specified number of normal distributions into a single vector.

Use of the function \code{\link{desireVar}} can be used to generate a mixture 
normal distribution with a specific global variance.
}
