% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{replicate_simulation_vary}
\alias{replicate_simulation_vary}
\title{Function to replication simulation with varying arguments}
\usage{
replicate_simulation_vary(sim_args)
}
\arguments{
\item{sim_args}{A nested named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}
}
\description{
This is a wrapper around \code{\link{simglm}} master function.
}
