% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{robust_model}
\alias{robust_model}
\title{Robust Model Standard Errors}
\usage{
robust_model(model, data = NULL, robust_args = list())
}
\arguments{
\item{model}{A fitted model object from lm or glm.}

\item{data}{A data object that contains a fitted model.}

\item{robust_args}{A named list of arguments passed to the robust model
standard error function.}
}
\description{
Robust Model Standard Errors
}
