% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.err.single}
\alias{sim.err.single}
\title{Function that simulates errors.}
\usage{
sim.err.single(errorVar, n, err.dist, num.dist)
}
\arguments{
\item{errorVar}{Scalar of error variance}

\item{n}{Cluster sample size.}

\item{err.dist}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{num.dist}{Number of distributions for bimod random variables.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}
\details{
Simulates error term for single level regression models.
}

