% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fixefSim.r
\name{sim.fixef.nested}
\alias{sim.fixef.nested}
\title{Simulates design matrix.}
\usage{
sim.fixef.nested(fixed, fixed.vars, cov.param, n, p, data.str,
  fact.vars = list(NULL))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.}

\item{fixed.vars}{Character vector of covariates for design matrix.}

\item{cov.param}{List of mean, sd (standard deviations), and var.type for fixed effects.
Does not include intercept, time, factors, or interactions.
var.type must be either "lvl1" or "lvl2". Must be same order as fixed formula above.}

\item{n}{Number of clusters.}

\item{p}{Number of within cluster units.}

\item{data.str}{Type of data. Must be "cross", or "long".}

\item{fact.vars}{A nested list of factor, categorical, or ordinal variable specification,
each list must include numlevels and var.type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}
}
\description{
Input fixed variables, sample size, and number of within variables, returns design matrix.
}
\details{
Simulates the fixed effects for the \code{\link{sim.reg}} function when a linear mixed
model is specified.  This function assumes a time variable when longitudinal data
is specified and does include any interactions that are specified.
}

