% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim.fixef.single}
\alias{sim.fixef.single}
\title{Simulates design matrix for single level model.}
\usage{
sim.fixef.single(fixed, fixed.vars, n, cov.param, fact.vars = list(NULL))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.}

\item{fixed.vars}{Character vector of covariates for design matrix.}

\item{n}{Number of clusters.}

\item{cov.param}{List of mean and sd (standard deviation) for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{fact.vars}{A list of factor, categorical, or ordinal variable specification, list must include
numlevels and var.type (must be "single" for single level regression);
optional specifications are: replace, prob, value.labels.}
}
\description{
Input fixed variables, sample size, and number of within variables, returns design matrix.
}
\details{
Simulates the fixed effects for the \code{\link{sim.reg}} function when simulating a
simple regression model.
}

