% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.pow.nested}
\alias{sim.pow.nested}
\title{Power simulation for nested designs}
\usage{
sim.pow.nested(fixed, random, fixed.param, random.param, w.var, cov.param, n, p,
  errorVar, randCor, rand.dist, err.dist, serCor, serCorVal, data.str,
  pow.param, alpha, pow.dist = c("z", "t"), pow.tail = c(1, 2))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{fixed.param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random.param}{Variance of random effects. Must be same length as random.}

\item{w.var}{Number of within cluster variables, including intercept if applicable.
Also could be number of level one covariates for cross-sectional clustering.}

\item{cov.param}{List of mean and variance for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{errorVar}{Scalar of error variance.}

\item{randCor}{Correlation between random effects.}

\item{rand.dist}{Simulated random effect distribution.  Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{err.dist}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{serCor}{Simulation of serial correlation. Must be "AR", "MA", "ARMA", or "ID", "ID" is default.}

\item{serCorVal}{Serial correlation parameters. A list of values to pass on to arima.sim.}

\item{data.str}{Type of data. Must be "cross", "long", or "single".}

\item{pow.param}{Number of parameter to calculate power includes intercept where applicable.}

\item{alpha}{What should the per test alpha rate be used for the hypothesis testing.}

\item{pow.dist}{Which distribution should be used when testing hypothesis test, z or t?}

\item{pow.tail}{One-tailed or two-tailed test?}
}
\description{
Takes simulation conditions as input, exports power.
}
\details{
Power function to compute power for a regression term for the linear mixed model.  This
function would need to be replicated to make any statement about power.  Use \code{\link{sim.pow}}
as a convenient wrapper for this.
}
\seealso{
\code{\link{sim.pow}} for a wrapper to replicate.
}

