% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/err_sim.r
\name{sim_err_nested}
\alias{sim_err_nested}
\title{Function that simulates errors.}
\usage{
sim_err_nested(error_var, n, p, rand_gen, arima = FALSE, ...)
}
\arguments{
\item{error_var}{Scalar of error variance}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{rand_gen}{The generating function used.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should
be correlated. If TRUE, must specify a valid model to pass to
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{...}{Additional specification needed to pass to the random generating
function defined by rand.gen.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}

