% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/err_sim.r
\name{sim_err_single}
\alias{sim_err_single}
\title{Function that simulates errors.}
\usage{
sim_err_single(error_var, n, rand_gen, arima = FALSE, ...)
}
\arguments{
\item{error_var}{Numeric scalar of error variance}

\item{n}{Cluster sample size.}

\item{rand_gen}{The generating function used.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should
be correlated. If TRUE, must specify a valid model to pass to
arima.sim, See \link{\code{arima.sim}} for examples.}

\item{...}{Additional values that need to be passed to the function
called from rand.gen.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}
\details{
Simulates error term for single level regression models.
}

