% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_fixef_nested}
\alias{sim_fixef_nested}
\title{Simulates design matrix.}
\usage{
sim_fixef_nested(fixed, fixed_vars, cov_param, n, p, data_str,
  cor_vars = NULL, fact_vars = list(NULL))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.}

\item{fixed_vars}{Character vector of covariates for design matrix.}

\item{cov_param}{List of mean, sd (standard deviations), and var_type for fixed effects. 
Does not include intercept, time, factors, or interactions. 
var_type must be either "lvl1" or "lvl2". Must be same order as fixed formula above.}

\item{n}{Number of clusters.}

\item{p}{Number of within cluster units.}

\item{data_str}{Type of data. Must be "cross", or "long".}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification, 
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}
}
\description{
Input fixed variables, sample size, and number of within variables, returns design matrix.
}
\details{
Simulates the fixed effects for the \code{\link{sim_reg}} function when a linear mixed
model is specified.  This function assumes a time variable when longitudinal data 
is specified and does include any interactions that are specified.
}

