% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_glm_func.r
\name{sim_glm_nested}
\alias{sim_glm_nested}
\title{Simulate two level logistic regression model}
\usage{
sim_glm_nested(fixed, random, fixed.param, random.param, cov.param, n, p,
  randCor, rand_dist, data_str, fact.vars = list(NULL), unbal = FALSE,
  unbalCont = NULL, ...)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{fixed.param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random.param}{Variance of random effects. Must be same length as random.}

\item{cov.param}{List of mean, sd (standard deviations), and var.type for fixed effects.
Does not include intercept, time, factors, or interactions.
var.type must be either "lvl1" or "lvl2". Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{randCor}{Correlation between random effects.}

\item{rand_dist}{Simulated random effect distribution.  Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{fact.vars}{A nested list of factor, categorical, or ordinal variable specification,
each list must include numlevels and var.type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{unbal}{A vector of sample sizes for the number of observations for each level 2
cluster. Must have same length as level two sample size n. Alternative specification
can be TRUE, which uses additional argument, unbalCont.}

\item{unbalCont}{When unbal = TRUE, this specifies the minimum and maximum level one size,
will be drawn from a random uniform distribution with min and max specified.}

\item{...}{Additional specification needed to pass to the random generating
function defined by rand.gen}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulates data for the nested logistic regression models.  Returns a
data frame with ID variables, fixed effects, random effects, and many
other variables to help when running simulation studies.
}
\examples{
\donttest{
fixed <- ~1 + time + diff + act + time:act
random <- ~1
fixed.param <- c(0.2, 1.5, 0.8, 1.2, 1.1)
random.param <- c(3)
cov.param <- list(mean = c(0, 0), sd = c(1.5, 4), var.type = c("lvl1", "lvl2"))
n <- 100
p <- 10
randCor <- 0
rand_dist <- "norm"
data_str <- "long"
temp.long <- sim_glm_nested(fixed, random, fixed.param, random.param,
 cov.param, n, p, randCor, rand_dist, data_str = data_str)
}
}

