% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{sim_pow_nested}
\alias{sim_pow_nested}
\title{Power simulation for nested designs}
\usage{
sim_pow_nested(fixed, random, fixed.param, random.param, w.var, cov.param, n, p,
  error_var, randCor, rand.dist, rand_gen, arima = FALSE, data.str, pow.param,
  alpha, pow.dist = c("z", "t"), pow.tail = c(1, 2))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{fixed.param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random.param}{Variance of random effects. Must be same length as random.}

\item{w.var}{Number of within cluster variables, including intercept if applicable.
Also could be number of level one covariates for cross-sectional clustering.}

\item{cov.param}{List of mean and variance for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{error_var}{Scalar of error variance.}

\item{randCor}{Correlation between random effects.}

\item{rand.dist}{Simulated random effect distribution.  Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{rand_gen}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should
be correlated. If TRUE, must specify a valid model to pass to
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{data.str}{Type of data. Must be "cross", "long", or "single".}

\item{pow.param}{Number of parameter to calculate power includes intercept where applicable.}

\item{alpha}{What should the per test alpha rate be used for the hypothesis testing.}

\item{pow.dist}{Which distribution should be used when testing hypothesis test, z or t?}

\item{pow.tail}{One-tailed or two-tailed test?}
}
\description{
Takes simulation conditions as input, exports power.
}
\details{
Power function to compute power for a regression term for the linear mixed model.  This
function would need to be replicated to make any statement about power.  Use \code{\link{sim_pow}}
as a convenient wrapper for this.
}
\seealso{
\code{\link{sim_pow}} for a wrapper to replicate.
}

