% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reg_func.r
\name{sim_reg_nested3}
\alias{sim_reg_nested3}
\title{Function to simulate three level nested data}
\usage{
sim_reg_nested3(fixed, random, random3, fixed_param, random_param = list(),
  random_param3 = list(), cov_param, k, n, p, error_var, with_err_gen,
  arima = FALSE, data_str, cor_vars = NULL, fact_vars = list(NULL),
  unbal = FALSE, unbal3 = FALSE, unbalCont = NULL, unbalCont3 = NULL,
  lvl1_err_params = NULL, arima_mod = list(NULL), contrasts = NULL, ...)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{random3}{One sided formula for random effects at third level in the simulation. Must be a subset of fixed
(and likely of random).}

\item{fixed_param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random_param}{A list of named elements that must contain: 
   random_var = variance of random parameters,
   rand_gen = Name of simulation function for random effects.
Optional elements are:
   ther: Theorectial mean and variance from rand_gen,
   ther_sim: Simulate mean/variance for standardization purposes,
   cor_vars: Correlation between random effects,
   ...: Additional parameters needed for rand_gen function.}

\item{random_param3}{A list of named elements that must contain: 
   random_var = variance of random parameters,
   rand_gen = Name of simulation function for random effects.
Optional elements are:
   ther: Theorectial mean and variance from rand_gen,
   ther_sim: Simulate mean/variance for standardization purposes,
   cor_vars: Correlation between random effects,
   ...: Additional parameters needed for rand_gen function.}

\item{cov_param}{List of arguments. Required arguments are:
 \itemize{
   \item dist_fun: This is a quoted R distribution function.
   \item var_type: This is the level of variable to generate. Must be 
     either 'lvl1', 'lvl2', or 'lvl3'. Must be same order as fixed formula 
     above.
 }
 Optional arguments to the distribution functions are in a nested list,
  see the examples for example code for this.
Does not include intercept, time, factors, or interactions.}

\item{k}{Number of third level clusters.}

\item{n}{Level two cluster sample size within each level three cluster.}

\item{p}{Within cluster sample size within each level two cluster.}

\item{error_var}{Scalar of error variance.}

\item{with_err_gen}{Simulated within cluster error distribution. Must be a quoted 'r' distribution
function.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should 
be correlated. If TRUE, must specify a valid model to pass to 
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification, 
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{unbal}{A vector of sample sizes for the number of observations for each level 2
cluster. Must have same length as level two sample size n. Alternative specification
can be TRUE, which uses additional argument, unbalCont.}

\item{unbal3}{A vector of sample sizes for the number of observations for each level 3
cluster. Must have same length as level two sample size k. Alternative specification
can be TRUE, which uses additional argument, unbalCont3.}

\item{unbalCont}{When unbal = TRUE, this specifies the minimum and maximum level one size,
will be drawn from a random uniform distribution with min and max specified.}

\item{unbalCont3}{When unbal3 = TRUE, this specifies the minimum and maximum level two size,
will be drawn from a random uniform distribution with min and max specified.}

\item{lvl1_err_params}{Additional parameters passed as a list on to the level one error generating function}

\item{arima_mod}{A list indicating the ARIMA model to pass to arima.sim. 
See \code{\link{arima.sim}} for examples.}

\item{contrasts}{An optional list that specifies the contrasts to be used for factor
variables (i.e. those variables with .f or .c). See \code{\link{contrasts}} for 
more detail.}

\item{...}{Not currently used.}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulates data for the linear mixed model, both cross sectional and longitudinal data.  Returns
a data frame with ID variables, fixed effects, and many other variables useful to help when 
running simulation studies.
}
\seealso{
\code{\link{sim_reg}} for a convenient wrapper for all data conditions.
}

