% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reg_func.r
\name{sim_reg_single}
\alias{sim_reg_single}
\title{Master function to simulate single level data.}
\usage{
sim_reg_single(fixed, fixed_param, cov_param, n, error_var, with_err_gen,
  arima = FALSE, data_str, cor_vars = NULL, fact_vars = list(NULL),
  lvl1_err_params = NULL, arima_mod = list(NULL), contrasts = NULL, ...)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{fixed_param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{cov_param}{List of arguments. Required arguments are:
 \itemize{
   \item dist_fun: This is a quoted R distribution function.
   \item var_type: This is the level of variable to generate. Must be 
     'single'. Must be same order as fixed formula above.
 }
 Optional arguments to the distribution functions are in a nested list,
  see the examples for example code for this.
Does not include intercept, factors, or interactions.}

\item{n}{Cluster sample size.}

\item{error_var}{Scalar of error variance.}

\item{with_err_gen}{Simulated within cluster error distribution. Must be a quoted 'r' distribution
function.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should 
be correlated. If TRUE, must specify a valid model to pass to 
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification, 
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{lvl1_err_params}{Additional parameters passed as a list on to the level one error generating function}

\item{arima_mod}{A list indicating the ARIMA model to pass to arima.sim. 
See \code{\link{arima.sim}} for examples.}

\item{contrasts}{An optional list that specifies the contrasts to be used for factor
variables (i.e. those variables with .f or .c). See \code{\link{contrasts}} for 
more detail.}

\item{...}{Not currently used.}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulates data for the simple regression models.  Returns a data frame with ID variables, 
fixed effects, and many other variables useful to help when running simulation studies.
}
\seealso{
\code{\link{sim_reg}} for a convenient wrapper for all data conditions.
}

