% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_time}
\alias{sim_time}
\title{Simulate Time}
\usage{
sim_time(n, time_levels = NULL, ...)
}
\arguments{
\item{n}{Sample size of the levels.}

\item{time_levels}{The values the time variable should take. If NULL (default),
the time values are discrete integers starting at 0 and going to n - 1.}
}
\description{
This function simulates data for the time variable of longitudinal data.
}
