% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simglm-package.r
\docType{package}
\name{simglm-package}
\alias{simglm-package}
\alias{_PACKAGE}
\title{simglm: Simulate Models Based on the Generalized Linear Model}
\description{
Simulates regression models, including both simple regression and generalized linear mixed models with up to three level of nesting. Power simulations that are flexible allowing the specification of missing data, unbalanced designs, and different random error distributions are built into the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lebebr01/simglm}
  \item Report bugs at \url{https://github.com/lebebr01/simglm/issues}
}

}
\author{
\strong{Maintainer}: Brandon LeBeau \email{lebebr01+simglm@gmail.com}

}
\keyword{internal}
