% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simglm_master_function.r
\name{simglm}
\alias{simglm}
\title{Single wrapper function}
\usage{
simglm(sim_args)
}
\arguments{
\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}
}
\description{
This function is most useful to pass to \code{\link{replicate_simulation}}.
  The function attempts to determine automatically which aspects to add to
  the simulation/power generation based on the elements found in the sim_args
  argument.
}
