% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{simulate_fixed}
\alias{simulate_fixed}
\title{Tidy fixed effect formula simulation}
\usage{
simulate_fixed(formula, n, contrasts = NULL)
}
\arguments{
\item{formula}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{n}{Sample size}

\item{contrasts}{Specification of the contrasts to be used 
for factor variables (i.e. those variables with .f/_f or .c/_c). 
See \code{\link{contrasts}} for more detail.}
}
\description{
This function simulates the fixed portion of the model using a formula syntax.
}
\examples{

}
