% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{simulate_knot}
\alias{simulate_knot}
\title{Simulate knot locations}
\usage{
simulate_knot(data, sim_args, ...)
}
\arguments{
\item{data}{Data to pass to the sim_knot2 function to determine knot 
location.}

\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{...}{Other arguments to pass to error simulation functions.}
}
\description{
Function that generates knot locations. An example of usefulness of this funciton
would be with generation of interrupted time series data. Another application may
be with simulation of piecewise linear data structures.
}
\examples{
sim_knot2()
}
