% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{compute_density_values}
\alias{compute_density_values}
\title{Convenience function for computing density values for plotting.}
\usage{
compute_density_values(data, group_var, parameter, values)
}
\arguments{
\item{data}{A dataframe that contains the parameter estimates.}

\item{group_var}{A group variable that specifies the attributes to
group by. By default, this would likely be the term attribute, but can
contain more than one attribute.}

\item{parameter}{The attribute that represents the parameter estimate.}

\item{values}{A list of numeric vectors that specifies the values
for which the density values are computed for.}
}
\description{
Convenience function for computing density values for plotting.
}
