% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glm.r
\name{data_glm_single}
\alias{data_glm_single}
\title{Generate logistic regression outcome}
\usage{
data_glm_single(Xmat, beta, n)
}
\arguments{
\item{Xmat}{A matrix of covariates.}

\item{beta}{A vector of regression parameters.}

\item{n}{Number of clusters.}
}
\description{
Takes simulation parameter arguments and 
returns simulated data in the logistic metric and
converted to 0/1 based on the probabilities and the 
binomial distribution with 1 trial.
}

