% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.r
\name{missing_data}
\alias{missing_data}
\title{Master Missing Data Function}
\usage{
missing_data(sim_data, resp_var = "sim_data", new_outcome = "sim_data2",
  clust_var = NULL, within_id = NULL, miss_prop = NULL,
  dropout_location = NULL, type = c("dropout", "random", "mar"), miss_cov)
}
\arguments{
\item{sim_data}{Simulated data frame}

\item{resp_var}{Character string of response variable with complete data.}

\item{new_outcome}{Character string of new outcome variable name that includes
the missing data.}

\item{clust_var}{Cluster variable used for the grouping, set to 
NULL by default which means no clustering.}

\item{within_id}{ID variable within each cluster.}

\item{miss_prop}{Proportion of missing data overall}

\item{dropout_location}{A vector the same length as the number of clusters 
representing the percentage of missing data for each cluster}

\item{type}{The type of missing data to generate, currently supports
droput, random, or missing at random (mar) missing data.}

\item{miss_cov}{Covariate that the missing values are based on.}
}
\description{
This function is a wrapper to easily call the specific types of 
missing data mechanisms.
}
