% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{replicate_simulation}
\alias{replicate_simulation}
\title{Replicate Simulation}
\usage{
replicate_simulation(sim_args, expression = NULL, ...)
}
\arguments{
\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{expression}{Simulation, model fitting, and coefficient extraction
expressions for a single replication.}

\item{...}{Currently not used.}
}
\description{
Replicate Simulation
}
