% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/errSim.r
\name{sim.err.nested}
\alias{sim.err.nested}
\title{Function that simulates errors.}
\usage{
sim.err.nested(errorVar, n, p, serCor, serCorVal, err.dist)
}
\arguments{
\item{errorVar}{Scalar of error variance}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{serCor}{Simulation of serial correlation. Must be "AR", "MA", "ARMA", or "ID", "ID" is default.}

\item{serCorVal}{Serial correlation parameters. A list of values to pass on to arima.sim.}

\item{err.dist}{Simulated within cluster error distribution. Must be "lap", "chi", "norm",
"norm" is default.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}

