% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim.err.nested}
\alias{sim.err.nested}
\title{Function that simulates errors.}
\usage{
sim.err.nested(errorVar, n, p, serCor, serCorVal, err.dist, num.dist, mean, var)
}
\arguments{
\item{errorVar}{Scalar of error variance}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{serCor}{Simulation of serial correlation. Must be "AR", "MA", "ARMA", or "ID", "ID" is default.}

\item{serCorVal}{Serial correlation parameters. A list of values to pass on to arima.sim.}

\item{err.dist}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{num.dist}{Number of distributions for bimod random variables}

\item{mean}{A vector of means passed to rbimod, length must match num.dist.}

\item{var}{A vector of variances passed to rbimod, length must match num.dist.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}

