% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.err.single}
\alias{sim.err.single}
\title{Function that simulates errors.}
\usage{
sim.err.single(errorVar, n, err.dist, num.dist, mean, var)
}
\arguments{
\item{errorVar}{Scalar of error variance}

\item{n}{Cluster sample size.}

\item{err.dist}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{num.dist}{Number of distributions for bimod random variables.}

\item{mean}{A vector of means passed to rbimod, length must match num.dist.}

\item{var}{A vector of variances passed to rbimod, length must match num.dist.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}
\details{
Simulates error term for single level regression models.
}

