% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errSim.r
\name{sim.err.single}
\alias{sim.err.single}
\title{Function that simulates errors.}
\usage{
sim.err.single(errorVar, n, err.dist, rand.gen = NULL, ...)
}
\arguments{
\item{errorVar}{Scalar of error variance}

\item{n}{Cluster sample size.}

\item{err.dist}{Simulated within cluster error distribution. Must be "lap", "chi", "norm",
"norm" is default.}

\item{rand.gen}{The generating function used with arima.sim.}

\item{...}{Additional values that need to be passed to the function
called from rand.gen.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}
\details{
Simulates error term for single level regression models.
}

