% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/errSim.r
\name{sim.err.single}
\alias{sim.err.single}
\title{Function that simulates errors.}
\usage{
sim.err.single(error_var, n, rand_gen, arima = FALSE, ...)
}
\arguments{
\item{error_var}{Numeric scalar of error variance}

\item{n}{Cluster sample size.}

\item{rand_gen}{The generating function used.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should
be correlated. If TRUE, must specify a valid model to pass to
arima.sim, \seealso{arima.sim}.}

\item{...}{Additional values that need to be passed to the function
called from rand.gen.}
}
\description{
Input error simulation parameters and outputs simulated errors.
}
\details{
Simulates error term for single level regression models.
}

