% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim.factor}
\alias{sim.factor}
\title{Simulate categorical, factor, or discrete variables}
\usage{
sim.factor(n, p, numlevels, replace = TRUE, prob = NULL,
  var.type = c("lvl1", "lvl2", "single"), value.labels = NULL)
}
\arguments{
\item{n}{Number of clusters or number of observations for single level}

\item{p}{Number of within cluster observations for multilevel}

\item{numlevels}{Scalar indicating the number of levels for categorical, factor, or discrete variable}

\item{replace}{Whether to replace levels of categorical variable, TRUE/FALSE}

\item{prob}{Probability of levels for variable, must be same length as numlevels}

\item{var.type}{Variable type for the variable, must be either "lvl1", "lvl2", or "single"}

\item{value.labels}{Optional argument with value labels for variable,
converts variable to factor.}
}
\description{
Function that simulates discrete, factor, or categorical variables.  Is essentially
a wrapper around the sample function from base R.
}

