% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_continuous}
\alias{sim_continuous}
\title{Simulate continuous variables}
\usage{
sim_continuous(k = NULL, n, p, dist, var_level = c(1, 2, 3), ...)
}
\arguments{
\item{k}{Number of third level clusters.}

\item{n}{Number of clusters or number of observations for single level}

\item{p}{Number of within cluster observations for multilevel}

\item{dist}{A distribution function. This argument takes a quoted
R distribution function (e.g. 'rnorm').}

\item{var_level}{The level the variable should be simulated at. This can either 
be 1, 2, or 3 specifying a level 1, level 2, or level 3 variable 
respectively.}

\item{...}{Additional parameters to pass to the dist_fun argument.}
}
\description{
Function that simulates continuous variables. Any distribution function in 
R is supported.
}
