% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_continuous}
\alias{sim_continuous}
\title{Simulate continuous variables}
\usage{
sim_continuous(k = NULL, n, p, dist_fun, var_type = c("lvl1", "lvl2",
  "lvl3", "single"), ...)
}
\arguments{
\item{k}{Number of third level clusters.}

\item{n}{Number of clusters or number of observations for single level}

\item{p}{Number of within cluster observations for multilevel}

\item{dist_fun}{A distribution function. This argument takes a quoted
R distribution function (e.g. 'rnorm').}

\item{var_type}{Variable type for the variable, must be either "lvl1", 
"lvl2", or "single"}

\item{...}{Additional parameters to pass to the dist_fun argument.}
}
\description{
Function that simulates continuous variables. Any distribution function in 
R is supported.
}

