% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_continuous2}
\alias{sim_continuous2}
\title{Simulate continuous variables}
\usage{
sim_continuous2(n, dist, ...)
}
\arguments{
\item{n}{Number of clusters or number of observations for single level}

\item{dist}{A distribution function. This argument takes a quoted
R distribution function (e.g. 'rnorm').}

\item{...}{Additional parameters to pass to the dist_fun argument.}
}
\description{
Function that simulates continuous variables. Any distribution function in 
R is supported.
}
