% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_factor}
\alias{sim_factor}
\title{Simulate categorical, factor, or discrete variables}
\usage{
sim_factor(k = NULL, n, p, numlevels, var_type = c("level1", "level2",
  "level3", "single"), ...)
}
\arguments{
\item{k}{Number of third level clusters.}

\item{n}{Number of clusters or number of observations for single level}

\item{p}{Number of within cluster observations for multilevel}

\item{...}{Additional parameters passed to the sample function.}

\item{levels}{Scalar indicating the number of levels for categorical, 
factor, or discrete variable. Can also specify levels as a character vector.}

\item{var_level}{The level the variable should be simulated at. This can either 
be 1, 2, or 3 specifying a level 1, level 2, or level 3 variable 
respectively.}

\item{replace}{TRUE/FALSE indicating whether levels should be sampled with 
replacement. Default is TRUE.}
}
\description{
Function that simulates discrete, factor, or categorical variables.  
Is essentially a wrapper around the sample function from base R.
}
