% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{sim_fixef_single}
\alias{sim_fixef_single}
\title{Simulates design matrix for single level model.}
\usage{
sim_fixef_single(fixed, fixed_vars, n, cov_param, cor_vars = NULL,
  fact_vars = list(NULL))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.}

\item{fixed_vars}{Character vector of covariates for design matrix.}

\item{n}{Number of clusters.}

\item{cov_param}{List of mean and sd (standard deviation) for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification,
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}
}
\description{
Input fixed variables, sample size, and number of within variables, returns design matrix.
}
\details{
Simulates the fixed effects for the \code{\link{sim_reg}} function when simulating a
simple regression model.
}

