% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_glm.R
\name{sim_glm}
\alias{sim_glm}
\title{Simulate quantities of interest from generalised linear models}
\usage{
sim_glm(obj, newdata, x_var, group_var, n = 1000, model = "lm", col_pal)
}
\description{
Simulate quantities of interest from generalised linear models
}
\examples{
# Normal Linear Model example
library(car) # Contains data
m1 <- lm(prestige ~ education + type,
         data = Prestige)

fitted_prestige <- expand.grid(education = 6:16, typewc = 1)

sim_glm(obj = m1, newdata = fitted_prestige, x_var = 'education')

fitted_prestige <- expand.grid(education = 6:16,
                               typewc = 0:1)
sim_glm(obj = m1, newdata = fitted_prestige, x_var = 'education',
        group_var = 'typewc')

# Logistic Model example
URL <- 'http://www.ats.ucla.edu/stat/data/binary.csv'
Admission <- read.csv(URL)
Admission$rank <- as.factor(Admission$rank)

m2 <- glm(admit ~ gre + gpa + rank,
          data = Admission, family = 'binomial')

fitted_admit_1 <- with(Admission,
                       expand.grid(gre = seq(220, 800, by = 10),
                                   gpa = mean(gpa),
                                   rank2 = 0:1))

sim_glm(obj = m2, newdata = fitted_admit_1, x_var = 'gre',
        group_var = 'rank2', model = 'logit')

fitted_admit_2 <- expand.grid(gre = seq(220, 800, by = 10), gpa = c(2, 3.5),
                              rank4 = 1)
sim_glm(obj = m2, newdata = fitted_admit_2, model = 'logit',
        x_var = 'gre', group_var = 'gpa')

}

