% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_reg.r
\name{sim_glm}
\alias{sim_glm}
\title{Master generalized simulation function.}
\usage{
sim_glm(fixed, random, random3, fixed_param, random_param = list(),
  random_param3 = list(), cov_param, k, n, p, data_str, cor_vars = NULL,
  fact_vars = list(NULL), unbal = FALSE, unbal3 = FALSE,
  unbalCont = NULL, unbalCont3 = NULL)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{random3}{One sided formula for random effects at third level in the simulation. Must be a subset of fixed
(and likely of random).}

\item{fixed_param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random_param}{A list of named elements that must contain: 
   random_var = variance of random parameters,
   rand_gen = Name of simulation function for random effects.
Optional elements are:
   ther: Theorectial mean and variance from rand_gen,
   ther_sim: Simulate mean/variance for standardization purposes,
   cor_vars: Correlation between random effects,
   ...: Additional parameters needed for rand_gen function.}

\item{random_param3}{A list of named elements that must contain: 
   random_var = variance of random parameters,
   rand_gen = Name of simulation function for random effects.
Optional elements are:
   ther: Theorectial mean and variance from rand_gen,
   ther_sim: Simulate mean/variance for standardization purposes,
   cor_vars: Correlation between random effects,
   ...: Additional parameters needed for rand_gen function.}

\item{cov_param}{List of mean and variance for fixed effects. Does not include intercept, time, or 
interactions. Must be same order as fixed formula above.}

\item{k}{Number of third level clusters.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification, 
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{unbal}{A vector of sample sizes for the number of observations for each level 2
cluster. Must have same length as level two sample size n. Alternative specification
can be TRUE, which uses additional argument, unbalCont.}

\item{unbal3}{A vector of sample sizes for the number of observations for each level 3
cluster. Must have same length as level two sample size k. Alternative specification
can be TRUE, which uses additional argument, unbalCont3.}

\item{unbalCont}{When unbal = TRUE, this specifies the minimum and maximum level one size,
will be drawn from a random uniform distribution with min and max specified.}

\item{unbalCont3}{When unbal3 = TRUE, this specifies the minimum and maximum level two size,
will be drawn from a random uniform distribution with min and max specified.}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulated data is useful for classroom demonstrations and to study 
the impacts of assumption violations on parameter estimates, statistical
power, or empirical type I error rates.

This function allows researchers a flexible approach to simulate regression
models, including single level models and cross sectional or longitudinal
linear mixed models (aka. hierarchical linear models or multilevel models).
}
\examples{
# generating parameters for single level regression
fixed <- ~1 + act + diff + numCourse + act:numCourse
fixed_param <- c(2, 4, 1, 3.5, 2)
cov_param <- list(mean = c(0, 0, 0), sd = c(4, 3, 3), var_type = c("single", "single", "single"))
n <- 150
temp.single <- sim_glm(fixed = fixed, fixed_param = fixed_param, cov_param = cov_param, 
n = n, data_str = "single")

# Longitudinal linear mixed model example
fixed <- ~1 + time + diff + act + time:act
random <- ~1 + time + diff
fixed_param <- c(4, 2, 6, 2.3, 7)
random_param <- list(random_var = c(7, 4, 2), rand_gen = 'rnorm')
cov_param <- list(mean = c(0, 0), sd = c(1.5, 4), var_type = c("lvl1", "lvl2"))
n <- 150
p <- 30
data_str <- "long"
temp.long <- sim_glm(fixed, random, random3 = NULL, fixed_param, random_param, random_param3 = NULL,
 cov_param, k = NULL, n, p, data_str = data_str)


# Three level example
fixed <- ~1 + time + diff + act + actClust + time:act
random <- ~1 + time + diff
random3 <- ~ 1 + time
fixed_param <- c(4, 2, 6, 2.3, 7, 0)
random_param <- list(random_var = c(7, 4, 2), rand_gen = 'rnorm')
random_param3 <- list(random_var = c(4, 2), rand_gen = 'rnorm')
cov_param <- list(mean = c(0, 0, 0), sd = c(1.5, 4, 2), 
var_type = c("lvl1", "lvl2", "lvl3"))
k <- 10
n <- 150
p <- 30
data_str <- "long"
temp.three <- sim_glm(fixed, random, random3, fixed_param, random_param, 
random_param3, cov_param, k,n, p, data_str = data_str)


}

