% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_glm_func.r
\name{sim_glm_nested}
\alias{sim_glm_nested}
\title{Simulate two level logistic regression model}
\usage{
sim_glm_nested(fixed, random, fixed_param, random_param = list(), cov_param,
  n, p, data_str, cor_vars = NULL, fact_vars = list(NULL), unbal = FALSE,
  unbalCont = NULL)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{fixed_param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random_param}{A list of named elements that must contain:
   random_var = variance of random parameters,
   rand_gen = Name of simulation function for random effects.
Optional elements are:
   ther: Theorectial mean and variance from rand_gen,
   ther_sim: Simulate mean/variance for standardization purposes,
   cor_vars: Correlation between random effects,
   ...: Additional parameters needed for rand_gen function.}

\item{cov_param}{List of mean, sd (standard deviations), and var_type for fixed effects.
Does not include intercept, time, factors, or interactions.
var_type must be either "lvl1" or "lvl2". Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification,
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{unbal}{A vector of sample sizes for the number of observations for each level 2
cluster. Must have same length as level two sample size n. Alternative specification
can be TRUE, which uses additional argument, unbalCont.}

\item{unbalCont}{When unbal = TRUE, this specifies the minimum and maximum level one size,
 will be drawn from a random uniform distribution with min and max specified.}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulates data for the nested logistic regression models.  Returns a
data frame with ID variables, fixed effects, random effects, and many
other variables to help when running simulation studies.
}
\examples{
\donttest{
fixed <- ~1 + time + diff + act + time:act
random <- ~1
fixed_param <- c(0.2, 1.5, 0.8, 1.2, 1.1)
random_param <- list(random_var = 3, rand_gen = 'rnorm')
cov_param <- list(mean = c(0, 0), sd = c(1.5, 4), var_type = c("lvl1", "lvl2"))
n <- 100
p <- 10
data_str <- "long"
temp.long <- sim_glm_nested(fixed, random, fixed_param, random_param,
 cov_param, n, p, rand_dist, data_str = data_str)
}
}

