% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_glm_func.r
\name{sim_glm_single}
\alias{sim_glm_single}
\title{Simulation single level logistic regression model}
\usage{
sim_glm_single(fixed, fixed.param, cov.param, n, data_str,
  fact.vars = list(NULL))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{fixed.param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{cov.param}{List of mean and sd (standard deviation) for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{fact.vars}{A nested list of factor, categorical, or ordinal variable specification,
     each list must include numlevels and var.type (must be "lvl1" or "lvl2");
     optional specifications are: replace, prob, value.labels.}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulates data for the simple logistic regression models.  Returns
a data frame with ID variables, fixed effects, and many other variables
to help when running simulation studies.
}
\examples{
\donttest{
# generating parameters for single level logistic regression
fixed <- ~1 + act + diff + numCourse + act:numCourse
fixed.param <- c(0.2, 1.5, 0.8, 1.2, 1.1)
cov.param <- list(mean = c(0, 0, 0), sd = c(1, 1, 1), var.type = c("single", "single", "single"))
n <- 150
temp.single <- sim_glm_single(fixed = fixed, fixed.param = fixed.param, cov.param = cov.param,
n = n)
# Fitting regression to obtain parameter estimates
summary(glm(sim.data ~ 1 + act + diff + numCourse + act:numCourse, data = temp.single,
family = "binomial"))

}
}

