% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{sim_pow_nested}
\alias{sim_pow_nested}
\title{Power simulation for nested designs}
\usage{
sim_pow_nested(fixed, random, fixed_param, random.param, cov_param, n, p,
  error_var, randCor, rand_dist, with_err_gen, arima, data_str, fact.vars,
  unbal, unbalCont, ..., pow_param, alpha, pow_dist = c("z", "t"),
  pow_tail = c(1, 2))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{fixed_param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random.param}{Variance of random effects. Must be same length as random.}

\item{cov_param}{List of mean, sd (standard deviations), and var.type for fixed effects.
Does not include intercept, time, factors, or interactions.
var.type must be either "lvl1" or "lvl2". Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{error_var}{Scalar of error variance.}

\item{randCor}{Correlation between random effects.}

\item{rand_dist}{Simulated random effect distribution.  Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{with_err_gen}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should
be correlated. If TRUE, must specify a valid model to pass to
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{fact.vars}{A nested list of factor, categorical, or ordinal variable specification,
each list must include numlevels and var.type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{unbal}{A vector of sample sizes for the number of observations for each level 2
cluster. Must have same length as level two sample size n. Alternative specification
can be TRUE, which uses additional argument, unbalCont.}

\item{unbalCont}{When unbal = TRUE, this specifies the minimum and maximum level one size,
will be drawn from a random uniform distribution with min and max specified.}

\item{...}{Additional specification needed to pass to the random generating
function defined by rand.gen.}

\item{pow_param}{Number of parameter to calculate power includes intercept where applicable.}

\item{alpha}{What should the per test alpha rate be used for the hypothesis testing.}

\item{pow_dist}{Which distribution should be used when testing hypothesis test, z or t?}

\item{pow_tail}{One-tailed or two-tailed test?}
}
\description{
Takes simulation conditions as input, exports power.
}
\details{
Power function to compute power for a regression term for the linear mixed model.  This
function would need to be replicated to make any statement about power.  Use \code{\link{sim_pow}}
as a convenient wrapper for this.
}
\seealso{
\code{\link{sim_pow}} for a wrapper to replicate.
}

