% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{sim_pow_nested}
\alias{sim_pow_nested}
\title{Power simulation for nested designs}
\usage{
sim_pow_nested(fixed, random, fixed_param, random_param = list(), cov_param,
  n, p, error_var, with_err_gen, arima = FALSE, data_str, cor_vars = NULL,
  fact_vars = list(NULL), unbal = FALSE, unbalCont = NULL,
  lvl1_err_params = NULL, arima_mod = list(NULL), missing = FALSE,
  missing_args = list(NULL), pow_param = NULL, alpha, pow_dist = c("z",
  "t"), pow_tail = c(1, 2), ...)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{random}{One sided formula for random effects in the simulation. Must be a subset of fixed.}

\item{fixed_param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{random_param}{A list of named elements that must contain: 
   random_var = variance of random parameters,
   rand_gen = Name of simulation function for random effects.
Optional elements are:
   ther: Theorectial mean and variance from rand_gen,
   ther_sim: Simulate mean/variance for standardization purposes,
   cor_vars: Correlation between random effects,
   ...: Additional parameters needed for rand_gen function.}

\item{cov_param}{List of mean, sd (standard deviations), and var_type for fixed effects. 
Does not include intercept, time, factors, or interactions. 
var_type must be either "lvl1" or "lvl2". Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{p}{Within cluster sample size.}

\item{error_var}{Scalar of error variance.}

\item{with_err_gen}{Simulated within cluster error distribution. Must be a quoted 'r' distribution
function.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should 
be correlated. If TRUE, must specify a valid model to pass to 
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{cor_vars}{A vector of correlations between variables.}

\item{fact_vars}{A nested list of factor, categorical, or ordinal variable specification, 
each list must include numlevels and var_type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{unbal}{A vector of sample sizes for the number of observations for each level 2
cluster. Must have same length as level two sample size n. Alternative specification
can be TRUE, which uses additional argument, unbalCont.}

\item{unbalCont}{When unbal = TRUE, this specifies the minimum and maximum level one size,
will be drawn from a random uniform distribution with min and max specified.}

\item{lvl1_err_params}{Additional parameters passed as a list on to the level one error generating function}

\item{arima_mod}{A list indicating the ARIMA model to pass to arima.sim. 
See \code{\link{arima.sim}} for examples.}

\item{missing}{TRUE/FALSE flag indicating whether missing data should be simulated.}

\item{missing_args}{Additional missing arguments to pass to the missing_data function. 
See \code{\link{missing_data}} for examples.}

\item{pow_param}{Name of variable to calculate power for, must be a name from fixed.}

\item{alpha}{What should the per test alpha rate be used for the hypothesis testing.}

\item{pow_dist}{Which distribution should be used when testing hypothesis test, z or t?}

\item{pow_tail}{One-tailed or two-tailed test?}

\item{...}{Not currently used.}
}
\description{
Takes simulation conditions as input, exports power.
}
\details{
Power function to compute power for a regression term for the linear mixed model.  This
function would need to be replicated to make any statement about power.  Use \code{\link{sim_pow}}
as a convenient wrapper for this.
}
\seealso{
\code{\link{sim_pow}} for a wrapper to replicate.
}

