% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pow_sim.r
\name{sim_pow_single}
\alias{sim_pow_single}
\title{Function to simulate power.}
\usage{
sim_pow_single(fixed, fixed.param, cov.param, n, error_var, rand_gen, pow.param,
  alpha, pow.dist = c("z", "t"), pow.tail = c(1, 2))
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{fixed.param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{cov.param}{List of mean and standard deviation for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{error_var}{Scalar of error variance.}

\item{rand_gen}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{pow.param}{Number of parameter to calculate power includes intercept where applicable.}

\item{alpha}{What should the per test alpha rate be used for the hypothesis testing.}

\item{pow.dist}{Which distribution should be used when testing hypothesis test, z or t?}

\item{pow.tail}{One-tailed or two-tailed test?}
}
\description{
Input simulation conditions and which term to compute power for, export reported power.
}
\details{
Power function to compute power for a regression term for simple regression models.  This
function would need to be replicated to make any statement about power.  Use \code{\link{sim_pow}}
as a convenient wrapper for this.
}
\seealso{
\code{\link{sim_pow}} for a wrapper to replicate.
}

