% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_reg_func.r
\name{sim_reg_single}
\alias{sim_reg_single}
\title{Master function to simulate single level data.}
\usage{
sim_reg_single(fixed, fixed.param, cov.param, n, error_var, rand_gen,
  arima = FALSE, data_str, fact.vars = list(NULL), ...)
}
\arguments{
\item{fixed}{One sided formula for fixed effects in the simulation.  To suppress intercept add -1 to formula.}

\item{fixed.param}{Fixed effect parameter values (i.e. beta weights).  Must be same length as fixed.}

\item{cov.param}{List of mean and sd (standard deviation) for fixed effects. Does not include intercept, time, or
interactions. Must be same order as fixed formula above.}

\item{n}{Cluster sample size.}

\item{error_var}{Scalar of error variance.}

\item{rand_gen}{Simulated within cluster error distribution. Must be "lap", "chi", "norm", "bimod",
"norm" is default.}

\item{arima}{TRUE/FALSE flag indicating whether residuals should
be correlated. If TRUE, must specify a valid model to pass to
arima.sim. See \code{\link{arima.sim}} for examples.}

\item{data_str}{Type of data. Must be "cross", "long", or "single".}

\item{fact.vars}{A nested list of factor, categorical, or ordinal variable specification,
each list must include numlevels and var.type (must be "lvl1" or "lvl2");
optional specifications are: replace, prob, value.labels.}

\item{...}{Additional specification needed to pass to the random generating
function defined by rand.gen.}
}
\description{
Takes simulation parameters as inputs and returns simulated data.
}
\details{
Simulates data for the simple regression models.  Returns a data frame with ID variables,
fixed effects, and many other variables useful to help when running simulation studies.
}
\seealso{
\code{\link{sim_reg}} for a convenient wrapper for all data conditions.
}

