% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simglm.r
\docType{package}
\name{simglm}
\alias{simglm}
\alias{simglm-package}
\title{simglm: A package to simulate and perform power by simulation for 
   models based on the generalized linear model.}
\description{
The simglm package provides two categories of important functions:
simulation functions (\code{\link{sim_reg}} and \code{\link{sim_glm}}) and 
power functions (\code{\link{sim_pow}} and \code{\link{sim_pow_glm}}).
#'
}
