% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simglm.r, R/simglm_master_function.r
\docType{package}
\name{simglm}
\alias{simglm}
\alias{simglm-package}
\alias{simglm}
\title{simglm: A package to simulate and perform power by simulation for 
   models based on the generalized linear model.}
\usage{
simglm(sim_args)
}
\arguments{
\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}
}
\description{
The simglm package provides two categories of important functions:
simulation functions (\code{\link{sim_reg}} and \code{\link{sim_glm}}) and 
power functions (\code{\link{sim_pow}} and \code{\link{sim_pow_glm}}).
#'

This function is most useful to pass to \code{\link{replicate_simulation}}.
  The function attempts to determine automatically which aspects to add to 
  the simulation/power generation based on the elements found in the sim_args
  argument.
}
