% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{simulate_fixed}
\alias{simulate_fixed}
\title{Tidy fixed effect formula simulation}
\usage{
simulate_fixed(data, sim_args, ...)
}
\arguments{
\item{data}{Data simulated from other functions to pass to this function. Can pass
NULL if first in simulation string.}

\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}

\item{...}{Other arguments to pass to error simulation functions.}
}
\description{
This function simulates the fixed portion of the model using a formula syntax.
}
\examples{

}
