% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixef_sim.r
\name{simulate_knot}
\alias{simulate_knot}
\title{Simulate knot locations}
\usage{
simulate_knot(data, sim_args)
}
\arguments{
\item{data}{Mostly internal argument.}

\item{sim_args}{A named list with special model formula syntax. See details and examples
for more information. The named list may contain the following:
\itemize{
  \item fixed: This is the fixed portion of the model (i.e. covariates)
  \item random: This is the random portion of the model (i.e. random effects)
  \item error: This is the error (i.e. residual term).
}}
}
\description{
Function that generates knot locations. An example of usefulness of this function
would be with generation of interrupted time series data. Another application may
be with simulation of piecewise linear data structures.
}
